requirements.txt
# Python requirements for development

# FastAPI & async
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg>=0.29.0

# Cache & Message Broker
redis==5.0.1
celery==5.3.4

# PDF Processing
pymupdf>=1.23.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
pyhanko>=0.20.0
pdf2image>=1.16.0

# Distributed locking
python-redlock>=2.2.0

# File handling
python-magic>=0.4.27
aiofiles>=23.0.0
pillow>=10.0.0

# Storage
minio>=7.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.8
brevo>=1.0.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client
requests>=2.31.0
httpx>=0.25.0

# File uploads
python-multipart>=0.0.6

# Antivirus
clamd>=1.0.2
pyclamd>=0.4.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.10.0
flake8>=6.0.0
mypy>=1.7.0
isort>=5.13.0
